requirements-api.txt
flask==2.1.0
flask-cors==3.0.10
flask-redoc==0.2.1
matplotlib==3.4.1
numpy
pandas==1.2.3
configargparse==1.5.2
python-dotenv==0.21.1
Flask-Cors==3.0.10
flask-login==0.6.0
flask-session==0.4.0
flask-wtf==1.0.1
Flask-Mail==0.9.1
WTForms==3.0.1
bs4==0.0.1
requests==2.26.0
mysql-connector-python==8.0.31
pyjwt==2.6.0
paramiko==3.3.1
Werkzeug==2.2.2